@@ -1,19 +1,15 @@
+import { RequestOptions } from '@angular/http';
+import { HttpClientModule, HttpClient } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { NgModule, LOCALE_ID } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpModule, Http, RequestOptions } from '@angular/http';
-import { AuthHttp, AuthConfig } from 'angular2-jwt';
-
 import { rootRouterConfig } from './app.routes';
-
 import { MyDatePickerModule } from 'mydatepicker';
-
 import { HttpApi } from './api/http-api.service';
 import { HttpRoomService } from './api/room/room-http.service';
 import { HttpAuthService } from './api/auth/auth-http.service';
-
 import { AppComponent } from './app.component';
 import { CalendarComponent } from './calendar/calendar.component';
 import { DatepickerComponent } from './calendar/datepicker/datepicker.component';
@@ -23,15 +19,12 @@ import { AdminComponent } from './admin/admin.component';
 import { CalendarDetailsComponent } from './calendar-details/calendar-details.component';
 import { HeaderComponent } from './header/header.component';
 import { LoadingModule, ANIMATION_TYPES } from 'ngx-loading';
+import { JwtModule } from '@auth0/angular-jwt';
 
-export function authHttpServiceFactory(http: Http, options: RequestOptions) {
-    return new AuthHttp(new AuthConfig(), http, options);
-}
 
 @NgModule({
   declarations: [
-      AppComponent,
-
+    AppComponent,
     CalendarComponent,
     DatepickerComponent,
     CreationReservationComponent,
@@ -41,11 +34,19 @@ export function authHttpServiceFactory(http: Http, options: RequestOptions) {
     HeaderComponent,
     
   ],
-  imports: [
+  imports: [  
     BrowserModule,
     FormsModule,
     ReactiveFormsModule,
-    HttpModule,   
+    HttpClientModule ,  
+    JwtModule.forRoot({
+      config: {
+        tokenGetter: () => {
+          return localStorage.getItem('token');
+        },
+        whitelistedDomains: ['localhost:4200', 'api.bureau401.fr']
+      }
+    }) , 
     RouterModule.forRoot(rootRouterConfig),	
     MyDatePickerModule,
     LoadingModule.forRoot({
@@ -60,9 +61,7 @@ export function authHttpServiceFactory(http: Http, options: RequestOptions) {
   providers: [
     HttpApi,
     HttpAuthService,
-    HttpRoomService,
-    { provide: LOCALE_ID, useValue: "fr-FR" },
-    { provide: AuthHttp, useFactory: authHttpServiceFactory, deps: [Http, RequestOptions]}
+    HttpRoomService,  
   ],
   bootstrap: [AppComponent]
 })
