@@ -1,27 +1,24 @@
+import { HttpApi } from './../http-api.service';
+import { HttpClient } from '@angular/common/http';
+
 import { Inject, Injectable, EventEmitter } from '@angular/core';
-import { Response, URLSearchParams, Headers } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
-import {JwtHelper} from 'angular2-jwt';
 import 'rxjs';
 
-import { HttpApi} from '../http-api.service';
+
 
 @Injectable()
-export class HttpAuthService  {
-    private jwtHelper: JwtHelper = new JwtHelper();
+export class HttpAuthService  {  
     private admin: boolean = false;
 
-    constructor(private http: HttpApi) {
-        let token = localStorage.getItem("token");
-        if (token != null){
-            this.admin = !this.jwtHelper.isTokenExpired(token);
-        }
+    constructor(private http: HttpClient, private HttpApi:HttpApi ) {
+       
     }
 
     public onLogin: EventEmitter<any> = new EventEmitter<any>();
 
-    loginAdmin(email: String, password: String): Observable<Response> {
-        let o = this.http.post('staff/login', { email: email, password: password }, null);
+    loginAdmin(email: String, password: String) {
+        let o = this.HttpApi.post('staff/login', {email:email,password:password},null );
         o.subscribe(data => {
              localStorage.setItem("token", data["token"]);
              this.admin = true;
@@ -30,8 +27,8 @@ export class HttpAuthService  {
         return o;
     }
 
-    loginClient(subscriberId: number, password: String): Observable<Response> {
-        let o = this.http.post('login', { id: subscriberId, password: password }, null);
+    loginClient(subscriberId: number, password: String) {
+        let o = this.http.post('login', { id: subscriberId, password: password } );
         o.subscribe(data => {
              localStorage.setItem("token", data["token"]);
              this.admin = false;
